In response to bugzilla bug #84, fixed all the Makefiles so that install-sh was found in the correct location.
The cause of this bug is that when configure converts foo.in => foo, it
also converts pathnames to relative pathnames - and according to
Makefile.header, install-sh would be ../install-sh.

The easiest way to get around this is just to have INSTALL_PROGRAM
be defined in Makefile.in and Makefile-plugins.in

Also, I removed the vaguely annoying '-fpic -fpic' bit, so '-fpic' is only
done once on the command line.