Oh, lots more junk.  The musicbrowser ui is getting functional.  Only things left to do i think are the playlist load/save operations and to slap the status bar at the bottom and start handling status events.
Wrote a metadata editor to go along with the playlist editor.  perty spiffy.
It stores stuff in the database.

Broke the build on win32, see utility.cpp.  win32 needs a FreeampDir() func to return the directory we're going to use for storage.  I'm not set on this way of
figuring out where to store our data, so ...

Moved ~/.freeamp_prefs to ~/.freeamp/preferences..  It should move the old
config file to the new position.